Add Google Analytics site tag
For now, add the GA script snippet directly to the header template.
We don't currently support plugins that add code to wp_head or
wp_footer, since we don't call those ourselves from our templates
in an effort to keep the site lightweight. (If we ever want to do
that instead, I have a prototype.)